+++ src/org/eclipse/swt/widgets/ColorDialog.d
@@ -147,7 +147,7 @@ public RGB open () {
         display.setModalDialog (this);
     }
     int signalId = 0;
-    int /*long*/ hookId = 0;
+    ptrdiff_t hookId = 0;
     CallbackData emissionData;
     emissionData.display = display;
     emissionData.data = handle;
